+++ example/lib/example_route.dart
@@ -21,6 +21,7 @@ import 'src/pages/simple/test_page_c.dart';
 FFRouteSettings getRouteSettings({
   required String name,
   Map<String, dynamic>? arguments,
+  Widget? notFoundWidget,
 }) {
   final Map<String, dynamic> safeArguments =
       arguments ?? const <String, dynamic>{};
@@ -149,6 +150,10 @@ FFRouteSettings getRouteSettings({
         routeName: 'MainPage',
       );
     default:
-      return const FFRouteSettings(name: '404', routeName: '404_page');
+      return FFRouteSettings(
+        name: '404',
+        routeName: '404_page',
+        widget: notFoundWidget,
+      );
   }
 }