+++ lib/src/template.dart
@@ -11,11 +11,11 @@ const String fileHeader = '''// GENERATED CODE - DO NOT MODIFY MANUALLY
 const String rootFile = """
 
 // ignore_for_file: prefer_const_literals_to_create_immutables,unused_local_variable,unused_import
-FFRouteSettings getRouteSettings({{1} String name, Map<String, dynamic>{2} arguments,}) {
+FFRouteSettings getRouteSettings({{1} String name, Map<String, dynamic>{2} arguments,Widget{2} notFoundWidget,}) {
   final Map<String, dynamic> safeArguments =arguments ?? const <String, dynamic>{};
   switch (name) {
 {0}   default:
-      return const FFRouteSettings(name:'404',routeName: '404_page');
+      return  FFRouteSettings(name:'404',routeName: '404_page',widget:notFoundWidget,);
   }
 }
 """;